Add initial working files.
Integrated.il contains the CIL to create a DLL with the new classes integrated. Original source of this version is not available.
The other .il files contain the complete CIL of all classes that were modified to create the mod.
The XML files contain data necessary to cause the new blocks to appear in game.